Fixed bug in energy exporter.

+++ src/org/wattdepot/hnei/csvimport/HneiImporter.java
@@ -12,17 +12,13 @@ import java.util.Collections;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
-import java.util.logging.FileHandler;
 import java.util.logging.Level;
 import java.util.logging.Logger;
-import java.util.logging.SimpleFormatter;
 import javax.xml.bind.JAXBException;
 import javax.xml.datatype.XMLGregorianCalendar;
-import org.wattdepot.client.OverwriteAttemptedException;
 import org.wattdepot.client.WattDepotClient;
 import org.wattdepot.client.WattDepotClientException;
 import org.wattdepot.datainput.RowParseException;
-import org.wattdepot.datainput.RowParser;
 import org.wattdepot.hnei.csvimport.validation.Entry;
 import org.wattdepot.hnei.csvimport.validation.EntrySortByTimestamp;
 import org.wattdepot.hnei.csvimport.validation.MonotonicallyIncreasingValue;
@@ -39,94 +35,7 @@ import au.com.bytecode.opencsv.CSVReader;
  * 
  * @author BJ Peter DeLaCruz
  */
-public class HneiImporter implements Importable {
-
-  /** Log file for this application. */
-  private static final Logger log = Logger.getLogger(HneiImporter.class.getName());
-
-  /** Output logging information to a text file. */
-  protected FileHandler txtFile;
-
-  /** Name of the file to be input. */
-  protected String filename;
-
-  /** URI of WattDepot server to send data to. */
-  protected String serverUri;
-
-  /** Name of Source to send data to. */
-  protected String sourceName;
-
-  /** Username to use when sending data to server. */
-  protected String username;
-
-  /** Password to use when sending data to server. */
-  protected String password;
-
-  /** Whether or not to skip the first row of the file. */
-  protected boolean skipFirstRow;
-
-  /** Name of the application on the command line. */
-  protected String toolName;
-
-  /** The parser used to turn rows into SensorData objects. */
-  protected RowParser parser;
-
-  /** Counts number of new sources added to the WattDepot server. */
-  protected int numNewSources;
-
-  /** Counts number of sources that are already on the WattDepot server. */
-  protected int numExistingSources;
-
-  /** Counts all sources that are in the CSV file. */
-  protected int numTotalSources;
-
-  /** Counts number of new data imported. */
-  protected int numNewData;
-
-  /** Counts number of data that already exists on the WattDepot server. */
-  protected int numExistingData;
-
-  /** Counts number of entries added to server. */
-  protected int numEntriesProcessed;
-
-  /** Counts number of entries that are invalid, e.g. do not contain any readings. */
-  protected int numInvalidEntries;
-
-  /** Counts total number of entries found in CSV file. */
-  protected int numTotalEntries;
-
-  /** Number of daily readings. */
-  protected int numDaily;
-
-  /** Number of hourly readings. */
-  protected int numHourly;
-
-  /** List of all entries in CSV file. */
-  protected List<Entry> entries;
-
-  /** List of all sources and their MTU IDs, with duplicates. */
-  protected List<Entry> allSources;
-
-  /** List of all sources and their MTU IDs, no duplicates. */
-  protected Set<Entry> allMtus;
-
-  /** List of all sources that have multiple MTU IDs. */
-  protected List<Entry> allDuplicateMtus;
-
-  /** List of all entries that have values that not monotonically increasing. */
-  protected List<Entry> allNonmonoIncrVals;
-
-  /** Start time in seconds to import data from CSV file. */
-  protected long importStartTime;
-
-  /** End time in seconds to import data from CSV file. */
-  protected long importEndTime;
-
-  /** Start time in seconds to validate data. */
-  protected long validateStartTime;
-
-  /** End time in seconds to validate data. */
-  protected long validateEndTime;
+public class HneiImporter extends Importer {
 
   /**
    * Creates a new HneiImporter object.
@@ -139,6 +48,7 @@ public class HneiImporter implements Importable {
    */
   public HneiImporter(String filename, String uri, String username, String password,
       boolean skipFirstRow) {
+    this.log = Logger.getLogger(HneiImporter.class.getName());
     this.filename = filename;
     this.serverUri = uri;
     this.sourceName = null;
@@ -169,12 +79,10 @@ public class HneiImporter implements Importable {
   }
 
   /**
-   * Returns the parser used to get rows from CSV files.
-   * 
-   * @return Parser used to get rows from CSV files.
+   * Sets the parser. Called after setting source name.
    */
-  public RowParser getParser() {
-    return this.parser;
+  public void setParser() {
+    this.parser = new HneiRowParser(this.toolName, this.serverUri, this.sourceName, log);
   }
 
   /**
@@ -184,151 +92,6 @@ public class HneiImporter implements Importable {
    */
   public void setSourceName(String sourceName) {
     this.sourceName = sourceName;
-  }
-
-  /**
-   * Sets the parser. Called after setting source name.
-   */
-  public void setParser() {
-    this.parser = new HneiRowParser(this.toolName, this.serverUri, this.sourceName, log);
-  }
-
-  /**
-   * Converts the runtime in milliseconds to the string format hh:mm:ss.
-   * 
-   * @param startTime Start time of a run.
-   * @param endTime End time of a run.
-   * @return Time in string format hh:mm:ss.
-   */
-  public String getRuntime(long startTime, long endTime) {
-    long milliseconds = endTime - startTime;
-    long hours = milliseconds / (1000 * 60 * 60);
-    long minutes = (milliseconds % (1000 * 60 * 60)) / (1000 * 60);
-    long seconds = ((milliseconds % (1000 * 60 * 60)) % (1000 * 60)) / 1000;
-    return String.format("%02d:%02d:%02d", hours, minutes, seconds);
-  }
-
-  /**
-   * Sets up the logger and file handler.
-   * 
-   * @return True if successful, false otherwise.
-   */
-  public boolean setupLogger() {
-    log.setLevel(Level.INFO);
-    try {
-      long timeInMillis = Calendar.getInstance().getTimeInMillis();
-      txtFile = new FileHandler(toolName + "-" + timeInMillis + ".log");
-      txtFile.setFormatter(new SimpleFormatter());
-    }
-    catch (IOException e) {
-      System.err.println("Unable to create file handler for logger.");
-      return false;
-    }
-    log.addHandler(txtFile);
-    return true;
-  }
-
-  /**
-   * Stores a source on the WattDepot server if it is not already on there.
-   * 
-   * @param client Used to store a source on the WattDepot server.
-   * @return True if successful, false otherwise.
-   */
-  public boolean storeSource(WattDepotClient client) {
-    this.sourceName = this.filename.substring(0, this.filename.lastIndexOf('.'));
-
-    // Store source on WattDepot server.
-    try {
-      client.storeSource(new Source(sourceName, username, true), false);
-    }
-    catch (OverwriteAttemptedException e) {
-      String msg = "Source " + sourceName + " already exists on server.\n";
-      System.out.print(msg);
-      log.log(Level.INFO, msg);
-    }
-    catch (WattDepotClientException e) {
-      System.err.println(e.toString());
-      log.log(Level.SEVERE, e.toString());
-      return false;
-    }
-    catch (JAXBException e) {
-      System.err.println(e.toString());
-      log.log(Level.SEVERE, e.toString());
-      return false;
-    }
-    return true;
-  }
-
-  /**
-   * Stores a source on a WattDepot server if it does not exist yet and then stores sensor data for
-   * that source.
-   * 
-   * @param client WattDepotClient used to connect to the WattDepot server.
-   * @param source Source that is described by the sensor data.
-   * @param datum Sensor data for a source.
-   * @return True if source and/or sensor data were stored successfully on WattDepot server.
-   */
-  public boolean process(WattDepotClient client, Source source, SensorData datum) {
-    try {
-      try {
-        client.storeSource(source, false);
-        this.numNewSources++;
-      }
-      catch (OverwriteAttemptedException e) {
-        this.numExistingSources++;
-        // log.log(Level.INFO, "Source " + source.getName() + " already exists on server.\n");
-      }
-      this.numTotalSources++;
-      client.storeSensorData(datum);
-      this.numNewData++;
-    }
-    catch (OverwriteAttemptedException e) {
-      this.numExistingData++;
-      String msg = "Data at " + datum.getTimestamp().toString() + " for " + source.getName();
-      msg += " already exists on server.\n";
-      log.log(Level.INFO, msg);
-    }
-    catch (WattDepotClientException e) {
-      System.err.println(e.toString());
-      log.log(Level.SEVERE, e.toString());
-      return false;
-    }
-    catch (JAXBException e) {
-      System.err.println(e.toString());
-      log.log(Level.SEVERE, e.toString());
-      return false;
-    }
-    return true;
-  }
-
-  /**
-   * Stores sensor data for a source that already exists on WattDepot server.
-   * 
-   * @param client WattDepotClient used to connect to the WattDepot server.
-   * @param datum Sensor data for a source.
-   * @return True if successful, false otherwise.
-   */
-  public boolean process(WattDepotClient client, SensorData datum) {
-    try {
-      client.storeSensorData(datum);
-      this.numNewData++;
-    }
-    catch (OverwriteAttemptedException e) {
-      this.numExistingData++;
-      String msg = "Data at " + datum.getTimestamp().toString() + " already exists on server.\n";
-      log.log(Level.INFO, msg);
-    }
-    catch (WattDepotClientException e) {
-      System.err.println(e.toString());
-      log.log(Level.SEVERE, e.toString());
-      return false;
-    }
-    catch (JAXBException e) {
-      System.err.println(e.toString());
-      log.log(Level.SEVERE, e.toString());
-      return false;
-    }
-    return true;
   }
 
   /**
@@ -358,7 +121,7 @@ public class HneiImporter implements Importable {
    * @param client WattDepotClient used to connect to the WattDepot server.
    * @param entry Current entry in CSV file.
    * @param datum SensorData for a source.
-   * @return Updated SensorData for a source.
+   * @return Updated SensorData object for a source.
    */
   public SensorData setSamplingInterval(WattDepotClient client, Entry entry, SensorData datum) {
     Calendar day = Calendar.getInstance();
@@ -384,6 +147,44 @@ public class HneiImporter implements Importable {
       datum.getProperties().getProperty().remove(new Property("daily", Boolean.toString(true)));
       this.numHourly++;
     }
+    return datum;
+  }
+
+  /**
+   * Adds energy and power data to a SensorData object.
+   * 
+   * @param client WattDepotClient used to connect to the WattDepot server.
+   * @param entry Current entry in CSV file.
+   * @param datum SensorData for a source.
+   * @return Updated SensorData object for a source.
+   */
+  public SensorData addProperties(WattDepotClient client, Entry entry, SensorData datum) {
+    List<SensorData> datas = null;
+    try {
+      XMLGregorianCalendar prevTimestamp = Tstamp.incrementDays(datum.getTimestamp(), -1);
+      datas = client.getSensorDatas(entry.getSourceName(), prevTimestamp, datum.getTimestamp());
+    }
+    catch (WattDepotClientException e) {
+      e.printStackTrace();
+      return null;
+    }
+
+    if (datas.size() > 2) {
+      double prevEnergy =
+          datas.get(datas.size() - 2).getPropertyAsDouble(SensorData.ENERGY_CONSUMED_TO_DATE);
+      double currEnergy = datum.getPropertyAsDouble(SensorData.ENERGY_CONSUMED_TO_DATE);
+      double energy = Math.abs(prevEnergy - currEnergy);
+      datum.addProperty(new Property(SensorData.ENERGY_CONSUMED, energy));
+
+      long prevTimestamp = datas.get(0).getTimestamp().toGregorianCalendar().getTimeInMillis();
+      long currTimestamp = datum.getTimestamp().toGregorianCalendar().getTimeInMillis();
+      double hours = Math.abs(prevTimestamp - currTimestamp) / 1000.0 / 60.0 / 60.0;
+      if (hours > 0) {
+        double power = energy / hours;
+        datum.addProperty(new Property(SensorData.POWER_CONSUMED, power));
+      }
+    }
+
     return datum;
   }
 
@@ -414,6 +215,47 @@ public class HneiImporter implements Importable {
     }
 
     Collections.sort(this.allDuplicateMtus);
+  }
+
+  /**
+   * Creates a CSV file that contains sources with multiple MTU IDs.
+   * 
+   * @return True if successful, false otherwise.
+   */
+  public boolean generateCsvFile() {
+    String today = Calendar.getInstance().getTime().toString().replaceAll("[ :]", "_");
+    File outputFile = new File(today + ".csv");
+    outputFile.setWritable(true);
+
+    BufferedWriter writer = null;
+    try {
+      writer = new BufferedWriter(new FileWriter(outputFile));
+    }
+    catch (IOException e) {
+      e.printStackTrace();
+      return false;
+    }
+
+    String str = null;
+    StringBuffer buffer = new StringBuffer();
+    str = "Account Number,MTU ID\n";
+    buffer.append(str);
+    for (Entry e : this.allDuplicateMtus) {
+      buffer.append(e.getSourceName());
+      str = "," + e.getMtuId() + "\n";
+      buffer.append(str);
+    }
+
+    try {
+      writer.write(buffer.toString());
+      writer.close();
+    }
+    catch (IOException e) {
+      e.printStackTrace();
+      return false;
+    }
+
+    return true;
   }
 
   /**
@@ -559,47 +401,6 @@ public class HneiImporter implements Importable {
   }
 
   /**
-   * Creates a CSV file that contains sources with multiple MTU IDs.
-   * 
-   * @return True if successful, false otherwise.
-   */
-  public boolean generateCsvFile() {
-    String today = Calendar.getInstance().getTime().toString().replaceAll("[ :]", "_");
-    File outputFile = new File(today + ".csv");
-    outputFile.setWritable(true);
-
-    BufferedWriter writer = null;
-    try {
-      writer = new BufferedWriter(new FileWriter(outputFile));
-    }
-    catch (IOException e) {
-      e.printStackTrace();
-      return false;
-    }
-
-    String str = null;
-    StringBuffer buffer = new StringBuffer();
-    str = "Account Number,MTU ID\n";
-    buffer.append(str);
-    for (Entry e : this.allDuplicateMtus) {
-      buffer.append(e.getSourceName());
-      str = "," + e.getMtuId() + "\n";
-      buffer.append(str);
-    }
-
-    try {
-      writer.write(buffer.toString());
-      writer.close();
-    }
-    catch (IOException e) {
-      e.printStackTrace();
-      return false;
-    }
-
-    return true;
-  }
-
-  /**
    * Parses each row, creates a SensorData object from each, and stores the sensor data on a
    * WattDepot server.
    * 
@@ -642,9 +443,9 @@ public class HneiImporter implements Importable {
       System.out.println("Reading in CSV file...\n");
 
       this.importStartTime = Calendar.getInstance().getTimeInMillis();
-      for (int i = 0; i < 100; i++) {
-        line = reader.readNext();
-      // while ((line = reader.readNext()) != null) {
+      // for (int i = 0; i < 100; i++) {
+      // line = reader.readNext();
+      while ((line = reader.readNext()) != null) {
         source = line[0];
         this.setSourceName(source);
         this.setParser();
@@ -657,7 +458,7 @@ public class HneiImporter implements Importable {
             entry.setMtuId(datum.getProperty("mtuID"));
             this.entries.add(entry);
 
-            if (this.process(client, new Source(source, username, true), datum)) {
+            if (this.process(client, new Source(source, this.username, true), datum)) {
               this.numEntriesProcessed++;
             }
             else {
@@ -702,12 +503,21 @@ public class HneiImporter implements Importable {
 
         // Classify data as either hourly or daily.
         datum = this.setSamplingInterval(client, e, datum);
+        if (datum == null) {
+          return false;
+        }
 
         temp = new Entry(e.getSourceName(), null, null, e.getMtuId());
         if (datum.getProperty("isMonotonicallyIncreasing").equals("false")) {
           temp.setMonotonicallyIncreasing(false);
         }
         this.allSources.add(temp);
+
+        // Add energyConsumed.
+        datum = this.addProperties(client, e, datum);
+        if (datum == null) {
+          return false;
+        }
 
         // Update sensor data on server.
         client.deleteSensorData(e.getSourceName(), e.getTimestamp());
@@ -760,8 +570,7 @@ public class HneiImporter implements Importable {
     String username = args[2];
     String password = args[3];
 
-    HneiImporter inputClient =
-      new HneiImporter(filename, serverUri, username, password, true);
+    HneiImporter inputClient = new HneiImporter(filename, serverUri, username, password, true);
 
     if (!inputClient.processCsvFile()) {
       System.exit(1);
